pyproject.toml
[project]
name = "botcategory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.20.0.post0",
    "apscheduler>=3.11.0",
    "asyncpg>=0.30.0",
    "dotenv>=0.9.9",
    "loguru>=0.7.3",
    "pytest-asyncio>=0.26.0",
    "sqlalchemy>=2.0.40",
]
